# src/services/connector-service/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Custos Connector Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/workspaces/{ws}/audit/leases": {
      "get": {
        "description": "List lease audit events for a workspace, newest-first.\n\nWhen ``eventType`` is omitted, the route fetches without an\nSPL-level event filter and post-filters to events whose\n``event_type`` is one of the known lease event types in\n:data:`_LEASE_EVENT_TYPES`. Callers narrowing to a single lease\nevent type SHOULD pass ``eventType`` so the SPL surface does the\nfilter and pagination stays dense.",
        "operationId": "list_lease_audit_events_v1_workspaces__ws__audit_leases_get",
        "parameters": [
          {
            "in": "path",
            "name": "ws",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ws",
              "type": "string"
            }
          },
          {
            "description": "Optional single event-type filter. Must be one of: lease.issued, lease.refreshed, lease.released, lease.expired, lease.revoke-requested, lease.revoked, lease.denied. Omit to query all lease events.",
            "in": "query",
            "name": "eventType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional single event-type filter. Must be one of: lease.issued, lease.refreshed, lease.released, lease.expired, lease.revoke-requested, lease.revoked, lease.denied. Omit to query all lease events.",
              "title": "Eventtype"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor"
            }
          },
          {
            "in": "query",
            "name": "occurredAfter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurredafter"
            }
          },
          {
            "in": "query",
            "name": "occurredBefore",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurredbefore"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 500,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query the audit log for lease-related events.",
        "tags": [
          "audit"
        ]
      }
    },
...
  }
}